Fix test failures in webos-qml test
