+++ packages/samples-loctool/Utils.js
@@ -18,7 +18,6 @@
  */
 
 const fs = require('fs');
-const xmljs = require("xml-js");
 
 const isValidPath = (filepath) => {
     return filepath ? fs.existsSync(filepath) : false;
@@ -39,16 +38,4 @@ const isExistKey = (filepath, key) => {
     return (jsonData && jsonData.hasOwnProperty(key));
 }
 
-const loadTSData = (filepath) => {
-    const options = {trim:false, nativeTypeAttribute: true, compact: true};
-
-    if (isValidPath(filepath)) {
-        const tsFile = fs.readFileSync(filepath, "utf-8");
-        if (tsFile) {
-            return xmljs.xml2js(tsFile, options).TS;
-        }
-    }
-    return undefined;
-}
-
-module.exports = { isValidPath, loadData, isExistKey, loadTSData };
+module.exports = { isValidPath, loadData, isExistKey };
